Open packages in the order the specification requires
The pieces existed separately; this is the sequence that binds them. The entry
count is refused before the archive is built, the media type is read at fixed
offsets before anything is parsed, and the version portal runs before any
judgement of validity.

That last ordering is the substance. A package from another era of the format
is not a broken package: 5.0 requires an unsupported version to be reported as
such, so a 0.8 file comes back as UnsupportedVersion with no violations, even
when it also contains faults this build would otherwise report.

Core XML goes through one loader, with DTD processing prohibited and no
resolver. A package is an untrusted file and a default XmlReader turns it into
a file-read primitive.

+++ src/Koma.Core/Packaging/ResourceLimits.cs
@@ -94,6 +94,6 @@ public sealed record ResourceLimits
         if (compressedBytes <= 0)
             return true;
 
-        return uncompressedBytes > compressedBytes * (long)MaxCompressionRatio;
+        return uncompressedBytes > compressedBytes * MaxCompressionRatio;
     }
 }
